+++ toggle-control.js
@@ -60,7 +60,6 @@ export default class ToggleControl extends GroundControl {
 
     this.pressed = this.pressed;
     this.initialValue = this.pressedValue || this.dataset.off;
-
     this.value = this.storedValue || this.initialValue;
   };
 
@@ -70,9 +69,8 @@ export default class ToggleControl extends GroundControl {
   };
 
   onValueChange = () => {
-    if (this.usedValue !== this.pressedValue) {
-      this.pressed = this.#toggleFromValue();
-    }
+    if (this.value === this.pressedValue) return;
+    this.pressed = this.#toggleFromValue();
   };
 
   doPress = (target) => {